fix: added phone number formatting, moved type to proper types folder, added phone number formatter
