+++ backend/package.json
@@ -68,6 +68,7 @@
     "geojson": "^0.5.0",
     "jest-mock": "^29.6.1",
     "jwks-rsa": "^3.0.1",
+    "libphonenumber-js": "^1.11.4",
     "linq-to-typescript": "^11.0.0",
     "nest-winston": "^1.9.2",
     "passport": "^0.6.0",
