+++ backend/src/types/models/reports/complaints/allegation-report-data.ts
@@ -1,0 +1,9 @@
+import { ComplaintReportData } from "./complaint-report-data";
+
+export interface AllegationReportData extends ComplaintReportData {
+  //-- ers
+  violationType: string;
+  inProgress: string;
+  wasObserved: string;
+  details: string;
+}
