+++ backend/src/types/models/reports/complaints/complaint-report-data.ts
@@ -1,0 +1,34 @@
+import { ComplaintUpdateDto } from "../../complaint-updates/complaint-update-dto";
+
+export interface ComplaintReportData {
+  reportDate: string;
+  reportTime: string;
+
+  id: string;
+  reportedOn: string;
+  updatedOn: string;
+  createdBy: string;
+  officerAssigned: string;
+  status: string;
+  incidentDateTime: string;
+  location: string;
+  latitude: string;
+  longitude: string;
+  community: string;
+  office: string;
+  zone: string;
+  region: string;
+  locationDescription: string;
+  description: string;
+
+  //-- caller information
+  name: string;
+  phone1: string;
+  phone2: string;
+  phone3: string;
+  email: string;
+  address: string;
+  reportedBy: string;
+
+  updates: Array<ComplaintUpdateDto>;
+}
