+++ backend/src/types/models/reports/complaints/wildlife-report-data.ts
@@ -1,0 +1,8 @@
+import { ComplaintReportData } from "./complaint-report-data";
+
+export interface WildlifeReportData extends ComplaintReportData {
+  //-- hwcr
+  natureOfComplaint: string;
+  species: string;
+  attractants: string;
+}
